front end v bookings

+++ app/controllers/cars_controller.rb
@@ -58,7 +58,7 @@ class CarsController < ApplicationController
   def destroy
     authorize @car
     @car.bookings.each do |booking|
-      booking.review.destroy
+      booking.review.destroy unless booking.review == nil
     end
     @car.bookings.destroy_all
     @car.destroy
@@ -72,6 +72,6 @@ class CarsController < ApplicationController
   end
 
   def car_params
-    params.require(:car).permit(:title, :content, :price, :address, :city, :photo)
+    params.require(:car).permit(:title, :content, :price, :address, :city, :photocar)
   end
 end
